ユーザ以外の置き換え
